add oracle miss count code
